Switch GridIndex to use float coordinates instead of int16s: fast way to prevent flicker from coordinate rounding.

+++ src/mbgl/text/collision_index.hpp
@@ -16,7 +16,7 @@ struct TileDistance;
 class CollisionIndex {
 public:
     using CollisionGrid = GridIndex<IndexedSubfeature>;
-    using GridUnit = int16_t;
+    using GridUnit = float;
 
     explicit CollisionIndex(const TransformState&);
